Refactoring to prepare for traction migration WIP
