+++ .gitignore
@@ -7,6 +7,7 @@
 
 # General
 **/*.env
+.DS_Store
 
 # Visual Studio Code
 .vscode
