+++ docker/docker-compose.yml
@@ -21,6 +21,13 @@ services:
       EMAIL_SUBJECT: ${EMAIL_SUBJECT}
       PUBLIC_SITE_URL: ${PUBLIC_SITE_URL}
       LOG_LEVEL: ${LOG_LEVEL}
+      ACAPY_AGENT_MODE: ${ACAPY_AGENT_MODE}
+      TRACTION_ENDPOINT: ${TRACTION_ENDPOINT}
+      TRACTION_INNKEEPER_PREFIX: ${TRACTION_INNKEEPER_PREFIX}
+      TRACTION_INNKEEPER_USER: ${TRACTION_INNKEEPER_USER}
+      TRACTION_INNKEEPER_PASSWORD: ${TRACTION_INNKEEPER_PASSWORD}
+      TRACTION_TENANT_PREFIX: ${TRACTION_TENANT_PREFIX}
+      TRACTION_ACAPY_WRAPPER_PREFIX: ${TRACTION_ACAPY_WRAPPER_PREFIX}
     volumes:
       - ../api:/usr/src/api
     ports:
@@ -51,6 +58,13 @@ services:
       PUBLIC_SITE_URL: ${PUBLIC_SITE_URL}
       LOG_LEVEL: ${LOG_LEVEL}
       S2I_SCRIPTS_PATH: ${S2I_SCRIPTS_PATH}
+      ACAPY_AGENT_MODE: ${ACAPY_AGENT_MODE}
+      TRACTION_ENDPOINT: ${TRACTION_ENDPOINT}
+      TRACTION_INNKEEPER_PREFIX: ${TRACTION_INNKEEPER_PREFIX}
+      TRACTION_INNKEEPER_USER: ${TRACTION_INNKEEPER_USER}
+      TRACTION_INNKEEPER_PASSWORD: ${TRACTION_INNKEEPER_PASSWORD}
+      TRACTION_TENANT_PREFIX: ${TRACTION_TENANT_PREFIX}
+      TRACTION_ACAPY_WRAPPER_PREFIX: ${TRACTION_ACAPY_WRAPPER_PREFIX}
     volumes:
       - ./api/config:/opt/app-root/src/config
     ports:
@@ -250,7 +264,7 @@ services:
     networks:
       - issuer_kit
     ports:
-      - ${WALLET_PORT}:${WALLET_PORT}
+      - ${WALLET_PORT_EXPOSED}:${WALLET_PORT}
     volumes:
       - wallet-data:/var/lib/postgresql/data
